pyproject.toml
[tool.poetry]
name = "streamlit_demo"
version = "0.1.0"
description = ""
authors = ["JJ Gong <jjgong7@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.10"
boto3 = "^1.24.49"
awswrangler = "^2.16.1"
streamlit = "^1.12.0"
pymongo = {extras = ["srv"], version = "^4.2.0"}
python-dotenv = "^0.20.0"
streamlit-aggrid = "^0.3.2"
pandas = "^1.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
